Switch to atomicfu in tests and sample project (#42)
This way, we can release the project (with passing CI and updated sample project) as soon as kotlinx.coroutines is published for major Kotlin updates– kotlinx.coroutines will always be published with an updated atomicfu. Aside from not having to wait on a dependency that isn't tied to coroutines, this doesn't provide much benefit over stately.